Add annotation to exclude containers from the readiness check
Katalog-sync works by syncing services to consul if the pod is "ready".
Until now there was a single exception for the sidecar, but with this
change you can exclude any container in the pod from this check. This
allows other side-car pods to control the readiness for k8s but not
influence the sync to consul