Make fetches from kubelet API non-fatal
This means if there is an issue talking to the kubelet API katalog-sync
will continue to enforce the last state it saw in k8s to consul. This
means you won't be able to add/remove things but it should effectively
"freeze" the state as it last saw it.

In the case that the daemon just started this will not be able to sync
state to consul at all since we won't have a starting state. In practice
this is no-worse since katalog-sync is run in-pod with consul-agent.