Check actual health of service on deregister call
Previously we would only allow the sidecar to be marked as deregistered
once the service was *missing* in consul. Unfortunately this meant that
sidecar services basically had to wait indefinitely; as the service
would exist in consul until it was removed in k8s -- which was waiting
on the sidecar. In this patch we start checking the service's health in
consul; if its not passing we consider it de-registered.